requirements.txt
amqp==2.5.2
asgiref==3.2.3
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.3.0
boto==2.49.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coverage==5.0.3
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
django-appconf==1.0.3
django-axes==5.2.2
django-braces==1.14.0
django-compressor==2.4
django-countries==6.0
django-crispy-forms==1.8.1
django-datatables-view==1.19.1
django-debug-toolbar==2.2
django-dynamic-formsets==0.0.8
django-extensions==2.2.8
django-formtools==2.1
django-grappelli==2.14.1
django-import-export==2.0.2
django-ipware==2.1.0
django-simple-history==2.8.0
django-sticky-messages==0.1
django-storages==1.9.1
django-taggit==1.2.0
django-timezone-field==4.0
django-wysiwyg-redactor==0.5.1
djangorestframework==3.11.0
docopt==0.4.0
et-xmlfile==1.0.1
fabric==2.5.0
factory-boy==2.12.0
Faker==4.0.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
invoke==1.4.1
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
kombu==4.6.7
mailchimp==2.0.9
MarkupPy==1.14
newrelic==5.8.0.136
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
paramiko==2.7.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycparser==2.19
pycrypto==2.6.1
Pygments==2.5.2
PyNaCl==1.3.0
PySocks==1.7.1
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
PyYAML==5.3
rcssmin==1.0.6
reportlab==3.5.34
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.14.0
South==1.0.2
sqlparse==0.3.0
tablib==1.1.0
text-unidecode==1.3
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
